fix(kernel,vfs,workspace): CI fixes + second review round for OS-level CoW
- kernel_router: extract workspace_commit_response (into response.rs) and the rate-limit metadata (into a new rate_limit.rs) so mod.rs stays under the 1000-line file-size cap; also restores rate_limit_max's doc adjacency.

- overlayfs: mount() is now an associated fn (drops the unused-&self clippy error); the mount mechanism is tracked in an AtomicU8 updated by BOTH prepare and rollback, since a re-mount may resolve via a different mechanism than the original (a OnceLock would go stale and mis-route unmount); unmount tries fusermount3 before fusermount (FUSE3 is the modern default).

- sandbox: each extra_masks entry is validated with validate_sandbox_str (absolute / UTF-8 / SBPL-safe) before use, matching the worktree and injection paths — masks are interpolated into the Seatbelt profile on macOS.

- CHANGELOG: drop the stale ancestor-walk git-managed bullet; the gitoxide entry (merged in #1169) is authoritative, and the non-git OS-level CoW entry stays.
