+++ CHANGELOG.md
@@ -24,7 +24,6 @@ Changelog tracking starts with 0.2.0. Prior versions were not tracked.
 ### Fixed
 
 - **A git-managed workspace no longer diverts capsule writes into the in-process copy-on-write overlay.** The workspace VFS was built as an `OverlayVfs` unconditionally at capsule load, so every capsule write landed in a per-capsule temporary upper directory instead of the real workspace — invisible to spawned `cargo`/tooling and to the user, with no commit path to reconcile it. When the workspace is inside a git work tree the overlay must not engage: git is already the rollback, and writes have to reach the real files that other processes read. Capsule load now detects git management with gitoxide work-tree discovery (`gix-discover`'s `upwards`, which walks up from `workspace_root` for an enclosing repository exactly as git does — correctly following the `.git` *file* a submodule/linked worktree uses, validating the discovered `.git`, and stopping at filesystem boundaries) and, when a work tree is found, backs `HostState.vfs` with a direct `HostVfs` registered at `workspace_root` (no overlay, no upper tempdir — `overlay_vfs`/`upper_dir` stay `None`, exactly like the existing plain-`HostVfs` lifecycle-hook state). Detection is upward-only — git has no robust "contains a repo below" primitive, and the agent workspaces this guards are themselves git roots — and fails safe to the overlay when no repository is found. A non-git workspace keeps today's overlay behaviour unchanged; detection is automatic with no config flag. The OS-sandbox writable root, path confinement, and the security gate are untouched — direct writes still target `workspace_root`, already the sandbox's writable bind.
-- **A git-managed workspace no longer diverts capsule writes into the in-process copy-on-write overlay.** The workspace VFS was built as an `OverlayVfs` unconditionally at capsule load, so every capsule write landed in a per-capsule temporary upper directory instead of the real workspace — invisible to spawned `cargo`/tooling and to the user, with no commit path to reconcile it. When the workspace is under (or itself contains) git version control the overlay must not engage: git is already the rollback, and writes have to reach the real files that other processes read. Capsule load now detects git management with a new bounded, cheap `astrid_core::dirs::workspace_is_git_managed(&Path)` helper (an ancestor walk for an enclosing work tree — matching a `.git` directory *or* the `.git` file a submodule/worktree uses — plus one shallow scan of the immediate child directories for a nested repo, skipping obvious noise trees) and, when true, backs `HostState.vfs` with a direct `HostVfs` registered at `workspace_root` (no overlay, no upper tempdir — `overlay_vfs`/`upper_dir` stay `None`, exactly like the existing plain-`HostVfs` lifecycle-hook state). A non-git workspace keeps today's overlay behaviour unchanged; detection is automatic with no config flag. The OS-sandbox writable root, path confinement, and the security gate are untouched — direct writes still target `workspace_root`, already the sandbox's writable bind.
 - **A non-git workspace now gets *real* OS-level copy-on-write, so spawned processes and the fs host share one filesystem.** The in-process `OverlayVfs` (the non-git branch of the fix above) diverted capsule writes into a tempdir upper that only the fs host could see; a spawned `cargo`/`git`/`build.rs` opened the workspace through the OS and read the pristine lower, never the upper, so copy-on-write was invisible to exactly the tools that matter. Capsule load now backs a non-git workspace with a new `astrid_vfs::workspace_cow` mechanism — a `WorkspaceCow` trait with an APFS `clonefile(2)` backend (macOS), an `overlayfs` mount backend (Linux, native `mount(2)` when the daemon holds mount authority, else `fuse-overlayfs`), and a fail-closed No-CoW backend — that establishes a single real **merged** directory the principal writes to AND spawned processes run in. `HostState.workspace_root` (the fs-host confinement root, the security gate's root, and the OS-sandbox writable root + cwd for every spawn) becomes that `merged_path`, so a spawned `cargo` builds against the copy-on-write tree where the writes actually are; `promote` (commit the merged tree into the pristine workspace) and `rollback` (discard it) are the safety net, exposed as audited `PromoteWorkspace`/`RollbackWorkspace` kernel admin requests (capabilities `self:workspace:promote` / `self:workspace:rollback`) and torn down on capsule unload. The copy-on-write upper/pristine directories are masked from spawned children (a new `extra_masks` set threaded into the sandbox's hidden-path list) so a child cannot write them directly and bypass the promote/rollback gate. When the clone/mount cannot be established the backend fails **closed** to direct writes (No-CoW) with a warning naming the reason — never a silently faked copy-on-write. The APFS and No-CoW backends are runtime-tested; the Linux `overlayfs` mount is compile-checked with an `#[ignore]`d CI-validated integration test. Making the workspace per-principal is a separate follow-up.
 
 ## [0.9.3] - 2026-07-06
